backend/pyproject.toml
[tool.poetry]
name = "mini-project"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
